pyproject.toml
[tool.poetry]
name = "sfc"
version = "0.1.0"
description = ""
authors = ["mojoee <moritzsontheimer@web.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
ipykernel = "^6.29.4"
jupyter = "^1.0.0"
langchain-community = "^0.2.1"
langchain-core = "^0.2.3"
langchain-experimental = "^0.0.59"
pypdf2 = "^3.0.1"
nltk = "^3.8.1"
opensearch-py = "^2.6.0"
python-dotenv = "^1.0.1"
transformers = "^4.41.2"
sentence-transformers = "^3.0.0"
pymilvus = "^2.4.4"
openai = "^1.37.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
